requirements.txt
oracledb
flask
python-dotenv
git-filter-repo
pandas
